+++ client/src/components/article/donut_graph.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import '../../stylesheets/donut_graph.css';
 import * as d3 from 'd3';
-
+import { visitPage } from '../wirr';
+let cheerio = require('cheerio')
+let URL = require('url-parse')
 
 // Inspired by : https://www.d3-graph-gallery.com/graph/donut_basic.html
               // https://medium.com/@kj_schmidt/show-data-on-mouse-over-with-d3-js-3bf598ff8fc2
@@ -9,17 +11,46 @@ export class DonutGraph extends React.Component {
     constructor(props) {
         super(props)
         this.state = {
-            data: { a: 10, b: 9, c: 14, d: 7, e: 12, f: 5, g: 19, h: 5, i: 10 },
+            data: {"loading...": 10, "please": 10, "wait": 10},
             margin: 40,
             width: 600,
             height: 450
         }
-
+        this.topTenAuthorContributionPercentage = this.topTenAuthorContributionPercentage.bind(this)
         this.drawChart = this.drawChart.bind(this)
     }
 
-    componentDidMount() {
+    async componentDidMount() {
+        this.topTenAuthorContributionPercentage(this.props.articleUrl)
         this.drawChart()
+    }
+
+    topTenAuthorContributionPercentage = (url) => {
+        // 1. Load under the 'url' package for proper parsing
+        const packagedUrl = new URL(url)
+        // 2. Get article name from pathanme by parsing
+        // (already joined by '_' from 'search.jsx')
+        const articleName = packagedUrl.pathname.split('/').slice(-1)[0]
+        // 3. Load it in the xtools wikipedia authorship statistics page
+        visitPage("https://xtools.wmflabs.org/authorship/en.wikipedia.org/" + articleName).then(res => {
+            const $2 = cheerio.load(res.body)
+
+            const authorsUsernames = $2('table.authorship-table td.sort-entry--username').slice(0, 10)
+                .map(function () { return $2(this).attr("data-value"); }).get()
+
+            const authorsContributionPercentage = $2('table.authorship-table td.sort-entry--percentage').slice(0, 10)
+                .map(function () { return $2(this).attr("data-value"); }).get()
+
+            const topTenAuthors = {};
+            authorsUsernames.forEach((author, idx) => topTenAuthors[author] = authorsContributionPercentage[idx]);
+
+            debugger
+            this.setState({
+                data: topTenAuthors
+            })   
+        }).catch(err => {
+            throw err
+        })
     }
 
     drawChart() {
